# Streamline Nova allocations (Arecibo backport)