--- a/src/r1cs/sparse.rs
+++ b/src/r1cs/sparse.rs
@@ -107,9 +107,23 @@ impl<F: PrimeField> SparseMatrix<F> {
   /// Multiply by a witness representing a dense vector; uses rayon to parallelize.
   /// This does not check that the shape of the matrix/vector are compatible.
   pub fn multiply_witness_unchecked(&self, W: &[F], u: &F, X: &[F]) -> Vec<F> {
-    let num_vars = W.len();
     // preallocate the result vector
-    let mut result = Vec::with_capacity(self.indptr.len() - 1);
+    let mut sink = Vec::with_capacity(self.indptr.len() - 1);
+    self.multiply_witness_into_unchecked(W, u, X, &mut sink);
+    sink
+  }
+
+  /// Multiply by a witness representing a dense vector; uses rayon to parallelize.
+  pub fn multiply_witness_into(&self, W: &[F], u: &F, X: &[F], sink: &mut Vec<F>) {
+    assert_eq!(self.cols, W.len() + X.len() + 1, "invalid shape");
+
+    self.multiply_witness_into_unchecked(W, u, X, sink);
+  }
+
+  /// Multiply by a witness representing a dense vector; uses rayon to parallelize.
+  /// This does not check that the shape of the matrix/vector are compatible.
+  pub fn multiply_witness_into_unchecked(&self, W: &[F], u: &F, X: &[F], sink: &mut Vec<F>) {
+    let num_vars = W.len();
     self
       .indptr
       .par_windows(2)
@@ -125,8 +139,7 @@ impl<F: PrimeField> SparseMatrix<F> {
             acc + val
           })
       })
-      .collect_into_vec(&mut result);
-    result
+      .collect_into_vec(sink);
   }
 
   /// number of non-zero entries