@srinathsetty 2023-12-06
Hi @huitseeker thanks for the PR!

> RecursiveSNARK-StepCircuitSize-0/Prove                       1.00     68.6±4.89ms             1.02    69.7±12.03ms
> RecursiveSNARK-StepCircuitSize-6565/Prove                    1.01     77.4±4.27ms             1.00     76.6±4.25ms
> RecursiveSNARK-StepCircuitSize-22949/Prove                   1.00    107.8±7.98ms             1.01   108.7±12.36ms
> RecursiveSNARK-StepCircuitSize-55717/Prove                   1.00    151.4±7.48ms             1.03   155.5±13.56ms
> RecursiveSNARK-StepCircuitSize-121253/Prove                  1.00    224.4±5.93ms             1.04   233.0±11.44ms
> RecursiveSNARK-StepCircuitSize-252325/Prove                  1.00   393.3±15.95ms             1.04   407.6±16.87ms
> RecursiveSNARK-StepCircuitSize-514469/Prove                  1.00   781.9±21.67ms             1.02   793.9±25.06ms
> RecursiveSNARK-StepCircuitSize-1038757/Prove                 1.00  1504.2±82.11ms             1.00  1500.8±37.38ms

From these lines, it seems like, these preallocations only save 4 milliseconds for the prover when the step circuit is of size a million gates. This is only 0.26% improvement. Am I reading this correctly?