@huitseeker 2023-12-06
> From these lines, it seems like, these preallocations only save 4 milliseconds for the prover when the step circuit is of size a million gates. This is only 0.26% improvement. Am I reading this correctly?

This is correct and expected. This benchmark does not stress the allocator.

This work was originally developed as a bundled package in https://github.com/lurk-lab/arecibo/pull/118 and tested in https://github.com/lurk-lab/lurk-rs/pull/890 what you're seeing here is broken-down changes.
In particular, for https://github.com/lurk-lab/arecibo/pull/139 we have benches at https://github.com/lurk-lab/lurk-rs/pull/923.