@porcuquine 2023-12-06
> > > From these lines, it seems like, these preallocations only save 4 milliseconds for the prover when the step circuit is of size a million gates. This is only 0.26% improvement. Am I reading this correctly?
> > 
> > 
> > This is correct and expected. This benchmark does not stress the allocator.
> 
> At what circuit size is this PR going to provide benefits? The PR proposes a fairly large, arguably somewhat complicated-to-maintain, change to the way memory is managed. I think it would be good to rigorously understand the benefits it provides. Specifically, for what circuits and what step circuit sizes is this beneficial?

Let me interpret the Lurk benchmarks shown here (you may need to click the disclosure triangle to see them. I will reproduce below):
https://github.com/lurk-lab/lurk-rs/pull/923#issuecomment-1827137943


Test | Base | PR | %
-- | -- | -- | --
LEM Fibonacci Prove - rc = 100/fib/num-100 | 4.8±0.03s | 4.4±0.03s | -8.33%
LEM Fibonacci Prove - rc = 100/fib/num-200 | 9.7±0.33s | 9.2±0.08s | -5.15%
LEM Fibonacci Prove - rc = 600/fib/num-100 | 3.8±0.01s | N/A | N/A
LEM Fibonacci Prove - rc = 600/fib/num-200 | 8.4±0.10s | N/A | N/A

The `rc` referenced is 'reduction count'. That is the number of inner Lurk circuits bundled in the application circuit. The Lurk circuit is about 12,000 constraints, so rc=100 is about 1.2M constraints. The fibonacci part (`num-100`, `num-200`) determines how large the calculation is, and we can derive the number of folding steps from that. In these cases, the calculation uses 11 and 33 steps respectively.

So, to synthesize, on Lurk's fibonacci workload, the answer to your question is:

For 1.2M constraints iterated 17 times, we see an 8.33% improvement.
For 1.2M constraints iterated 33 times, we see a 5.15% improvement.

Note that these metrics are a bit tricky to interpret because the first step is cheaper and therefore biases toward better performance on shorter computations.

It's possible I presented some of this info wrong, but that is the basic answer. @winston-h-zhang can fill in or fix any details I omitted or got wrong.