@huitseeker 2023-12-07
For more signal on memory, here is the top-level memory profiling of the last round of the minroot example (65536 it/step) using [`bytehound`](https://github.com/koute/bytehound). First is on main, second is the PR (commit hashes top center in the picture).  

<img width="1753" alt="On main" src="https://github.com/microsoft/Nova/assets/4142/65fd42b4-50ca-4807-9365-92d7169a3c00">
<img width="1762" alt="On PR" src="https://github.com/microsoft/Nova/assets/4142/d9f22098-bcf0-43ec-aaf6-d1c45af668d2">

As you can see, the overall memory usage is far from taxing for a modern machine. But it's sufficient to see the nine steps of folding that each create their short-lived allocation. 

As to the extrapolation of how this leads to a degradation in performance through memory fragmentation, here's what happens if if zoom out of the first picture (on main) to see all 7 rounds of Minroot:
<img width="1779" alt="Screenshot 2023-12-06 at 8 53 22 PM" src="https://github.com/microsoft/Nova/assets/4142/97e08675-7aa2-4924-a486-ca0325b60d4a">
As you can see, the memory usage pattern scales linearly with the number of constraints.