@winston-h-zhang 2023-12-07
To elaborate more on @porcuquine and @huitseeker, I've provided some more benchmarks that give more context to why we've pursued this new memory model, and the breakdowns of improvements each PR provides.

## Context

In theory, having many large reallocations of vectors puts a lot of load on the memory allocator, due to factors like fragmentation. Thus, we should try to avoid reallocating memory whenever it is possible to reuse already available space. Concretely, the following benchmark compares running the proving pipeline under (1) ideal memory conditions and (2) unexpectedly bad memory conditions.

| rc  | Test           | Base (fresh)  | Base (load)      | %          |
|-----|----------------|---------------|------------------|------------|
| 100 | Fibonacci(100) | **6.60s**     | 7.47s            | **113%**  |
|     | Fibonacci(200) | **14.58s**    | 14.89s           | **102%**  |
| 600 | Fibonacci(100) | **4.33s**     | 6.34s            | **146%**  |
|     | Fibonacci(200) | **10.24s**    | 14.65s           | **143%**  |

In this particular situation, "fresh/load" indicates whether the Lurk program generates fresh public parameters and loads them from memory. Choosing fresh vs load has no semantic difference to how Lurk runs. However, the result is a huge regression that seems to have come from nowhere. More generally, we can imagine a scenario where a change in some variable of the system which makes no semantic difference to the proving pipeline, but somehow causes a regression. Because the current memory pipeline is unstable, any change could **potentially** mess with the allocator and cause regressions, while remaining extremely undetectable to debug.